+++ src/pages/Category/EditModal.tsx
@@ -4,7 +4,7 @@ import { Form, Input, Modal, TreeSelect, Typography, Divider } from 'antd'
 import { DataNode } from 'antd/lib/tree'
 import { FormDataExtend } from '@/components/common/FormDataExtend'
 import { Category as CategoryType } from '@/constants/category'
-import { stringToYMD } from '@/transformers/date'
+import { stringToYMD } from '@/transforms/date'
 
 const CATEGORY_NULL_VALUE = null as any
 const formLayout = {