+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+### v1.2.3 (2021-12-10)
+
+- Add `MultipleImageUploader` component
+- Rename `transformers` to `transforms`
+
 ### v1.2.2 (2021-12-07)
 
 - Fix `per_page` value to `50`